Refactor roamer systems to anomaly namespace
Moved roamer spawning and movement logic from the entity namespace to the anomaly namespace. Introduced AnomalyRoamer component and AnomalyState struct for improved state management. Updated GameState and system update calls accordingly. Added vector normalization utilities and removed deprecated roamer entity code.

+++ src/systems/object/receiver/artillery_control_system.cpp
@@ -106,10 +106,10 @@ ArtilleryControlSystem::CoordResult ArtilleryControlSystem::InterpretMessageAsCo
 	switch (result.axis)
 	{
 	case CoordResult::Horizontal:
-		regionLength = RenderContext::RADAR_SIZE.x;
+		regionLength = static_cast<float>(RenderContext::RADAR_SIZE.x);
 		break;
 	case CoordResult::Vertical:
-		regionLength = RenderContext::RADAR_SIZE.y;
+		regionLength = static_cast<float>(RenderContext::RADAR_SIZE.y);
 		break;
 	default:
 		break;
